+++ src/Identity.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\HttpAuthentication;
-
-interface Identity
-{
-    public function toString(): string;
-}