+++ src/ViaUrlAuthority.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\HttpAuthentication;
 
-use Innmind\HttpAuthentication\ViaUrlAuthority\Resolver;
 use Innmind\Http\ServerRequest;
 use Innmind\Url\Authority\UserInformation\{
     User,
@@ -11,22 +10,32 @@ use Innmind\Url\Authority\UserInformation\{
 };
 use Innmind\Immutable\Attempt;
 
-final class ViaUrlAuthority implements Authenticator
+/**
+ * @template T
+ */
+final class ViaUrlAuthority
 {
-    private Resolver $resolve;
+    /** @var callable(User, Password): Attempt<T> */
+    private $resolve;
 
-    public function __construct(Resolver $resolve)
+    /**
+     * @param callable(User, Password): Attempt<T> $resolve
+     */
+    public function __construct(callable $resolve)
     {
         $this->resolve = $resolve;
     }
 
+    /**
+     * @return Attempt<T>
+     */
     public function __invoke(ServerRequest $request): Attempt
     {
         $user = $request->url()->authority()->userInformation()->user();
         $password = $request->url()->authority()->userInformation()->password();
 
         if ($user->equals(User::none()) && $password->equals(Password::none())) {
-            /** @var Attempt<Identity> */
+            /** @var Attempt<T> */
             return Attempt::error(new \RuntimeException('No authentication provided'));
         }